Some more pages added.

+++ src/main/java/com/ecommerce/year2_sem2_project/Entity/Product.java
@@ -18,7 +18,7 @@ public class Product {
     private String description;
 
     @Column(name = "price")
-    private BigDecimal price;
+    private Long price;
 
     @Column(name = "stock")
     private Integer stock;
@@ -50,11 +50,11 @@ public class Product {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public Long getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Long price) {
         this.price = price;
     }
